Made route Private for CRUD operations

+++ controllers/contactControllers.js
@@ -7,15 +7,15 @@ const Contact = require("../models/contactModal");
 
 // @desc Get all contacts
 // @routes GET /api/contacts
-// @access public
+// @access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 // @desc Create New Contact
 // @routes POST /api/contacts
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res) => {
   console.log("The req body for POST is -", req.body);
   const { name, email, phone } = req.body;
@@ -27,16 +27,21 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json(contact);
 });
 
 // @desc Get Contact
 // @routes GET /api/contacts/:id
-// @access public
+// @access private
 const getContact = asyncHandler(async (req, res) => {
   try {
-    const contact = await Contact.findById(req.params.id);
+    // const contact = await Contact.findById(req.params.id);
+    const contact = await Contact.findOne({
+      _id: req.params.id, // Match the contact's own id
+      user_id: req.user.id, // Match the user_id
+    });
     console.log("error", contact, req.params.id);
     if (!contact) {
       res.status(404); // Set the status code to 404
@@ -51,7 +56,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 // @desc Update Contact
 // @routes PUT /api/contacts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler(async (req, res) => {
   try {
     const contact = await Contact.findById(req.params.id);
@@ -60,6 +65,12 @@ const updateContact = asyncHandler(async (req, res) => {
       res.status(404); // Set the status code to 404
       throw new Error("Contact not found");
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+      res.status(403);
+      throw new Error("User dont have permission to update");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
       req.params.id,
       req.body,
@@ -75,7 +86,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 // @desc Delete Contact
 // @routes DELETE /api/contacts/:id
-// @access public
+// @access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   console.log("error", contact, req.params.id);
@@ -83,6 +94,12 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404); // Set the status code to 404
     throw new Error("Contact not found");
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User dont have permission to delete");
+  }
+
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
 });
